Debugging fork() syscall
- There was bugs in syscall.c so proc syscalls were not called at all
- Fixing stupid bugs here and there...

fork() can be now properly called but it seems to be very unstable and
the syscall seems to cause a kernel panic during the fork or few
millisecs after fork() exits. Reasons are unknown but it seems that
something tries to touch vregions during or just after fork() and
it might be that the actual fault is triggered by COW procedure.