Fixed FB logger
There is a bit critical issue if FB logger is set as a default klogger
because kerror can't have a dependency to actual fb drivers and fb
abstraction doesn't know if there will be any FBs registered ever. So
essentially this causes kerror to write to FB too early before
initialization is ready.

Possible solution for this kind of problems might be to add a "querying"
interface to harware abstraction subsystems that would try to init all
drivers in a known linker set. Then other modules could have true
dependecy to the abstraction level. This of course doesn't remove the
problem with issues if still no devices have been actually registered.