Fixing weird crashes
So idle thread stack was way too small and kernel crashed if it got into
too many maintenance tasks in idle thread as its stack wiped some of the
scheduler's data. Actual solution was to increase the stack size and to
make it configurable via Kconfig but this change fixes lot of minor bugs
that may also cause strange behavior.