Fixing NULL pointer issues with ramfs
Hunting down a prefetch abort caused by a failure in ramfs.
It seems that ramfs will sometimes/always fail if a directory is created
and then created again with the same name. The resulting directory is
only partially created, an inode pointer is hashed but it seems to be
NULL which confuses get_vnode part of the code.

This commit adds a lot of asserts to check if a vnode or inode is a NULL
pointer thus limiting the damage caused by a broken ramfs. We still have
to find out how to detect the failure before mkdir exits and finally
fix the actual bug.

This commit also hardens the kernel by disabling execution on tkstack
which caused a big mess after vn->vn_ops->fn() was called in procfs.