Add a giant lock to kmalloc
kmalloc is very unsafe for a preemptible kernel at the moment, this
commit tries to fix this issue by introducing atomic ref counts and
a giant lock for operations. However as a side effect this commit
will introduce some new dead locks because kmalloc is used in some
interrupt handlers, even though no memory allocation functions should
be used in interrupt handlers. That issue shall be covered in a
separate commit.