Ensure that a signal queued is always forwarded
If a signal is delivered to a process or a thread and
the receiving thread is in sleep for any reason the
signal handling is some times postponed even if it
could start right away ie. the thread is not blocked.

The previous is seems to be a problem for wait()
syscalls since the main thread is sleeping and it
can't receive a signal 7 until it's too late and
thus wastes ̃~1 second.

This commit changes the signal queuing behavior in
a way that it always tries to wake up the receiving
thread for signal forwarding if possible.