Break thread inherintance connection between processes
Since death of a parent thread causes killing of its children, it's not
a desirable feature to have inheritance for all threads across process
container boundaries. Killing child threads causes also child processes
to be killed which is against the idea of a life cycle of a process in
POSIX.

By having a NULL parent for main() thread we get rid of killing of the
children and it also permits us to have daemons in a traditional way.