Avoid use of thread_terminate()
Use of thread_terminate() will cause resources beign left as is
when the thread is terminated which is not very nice. To avoid this
prolem we should remove syscalls calling it as well as make sure
thread_terminate() is not used as the normal way of killing a thread,
instead ksignaling should be used because it tries to kill threads
a bit more gracefully.

Issue #117