Fix vnode ref bug in lookup
The following lookup caused ramfs based file systems to
run to a negative number of references to its root.
The fix is to never vref() if a fs lookup function returns
-EDOM and when the caller of the lookup function knows the
special meaning of -EDOM it should respectively take vref().

This commit also optmizes mount point travelsals a bit and
fixes a possibly ref in there as well.

cd proc # to a different fs than /
ls ../mnt # lookup to a third type