Revamp process capabilities and make it a mandatory feature
In the future process capabilities will be the only way to manage
process privileges and uid=0 will have no special meaning inside
the kernel.

- Remove `#ifdefs` and Kconfig that allows disabling the
  capability system
- Revamp the capability code allocation
- Remove grant and restricted capabilities completely as those
  make very little sense in a system that has no notion of a root
  user
- Creat a new simple capability system based on an effective and
  bounding set of capabilities resticting the process privileges

Currently this system doesn't yet allow us to remove the special
root user because there is no way to grant capabilities per
executable and there is no point trying to handle the capabilities
in the system daemons.

Handling File System Capabilities
---------------------------------

Create a new capability that allows/disallows opening a file for
write that has the FAT `system` attribute set. This will give a
neat additional protection for system files.

Now add a new `elf` header specifying the required capabilities.
The capabilities can be only assigned if the execing process
has the capabilities in the bounding set. A capability allowing
to modify effective capabilities is not needed. The header is
only processed if the `system` attribute is set.