Remove broken Elf32_Syminfo typedef
`Elf32_Syminfo` is typically a typedef but we normally
avoid typedefs in Zeke and this is possibly a some sort
of leftover. The implementation of the elf headers isn't
compliant with many other OSes, but this is state doesn't
change much if this broken typedef, now a global variable,
is removed.